cleanup(angular-rspack): reorganize the engine manifest internals and version probing helpers
Move the engine manifest source generation off the exported loader
subpath into an internal module, rename getInstalledPackageVersion to
getInstalledPackageVersionFromRoot to avoid colliding with the
@nx/devkit/internal helper of the same name, and narrow its catch to
module-resolution errors so unexpected failures surface. Also mark the
rspack virtual-modules probe with a TODO(v24), correct a stale comment in
the server assets, and pin the 1.5.0 probe boundary with a test.
